feat(AdminController): Tính toán dữ liệu Dashboard (tiền trong ngày, trong tháng, người tạo tài khoản trong ngày, trong tháng) ver 1.0 học thêm: base record, indexing, hash<T>

+++ CRUD_asp.netMVC/Service/Payment/QrCodeService.cs
@@ -2,7 +2,7 @@
 {
     public class QrCodeService
     {
-        public string GenerateBankQrCode(string accountNo, double? amount, string description, string accountName)
+        public string GenerateBankQrCode(string accountNo, decimal? amount, string description, string accountName)
         {
             // Su dung vietQR 
             var bankBin = "970422"; // Mã bin MB Bank
